buildifier_test: fix `FINDSTR: Cannot open MANIFEST`
This happens on Windows when using `buildifier_test`:
```
INFO: Running command line: 'external\bazel_tools\tools\test\tw.exe' buildifier/test.bat
Executing tests from //buildifier:test
-----------------------------------------------------------------------------
FINDSTR: Cannot open MANIFEST
The expression after '&' in a pipeline element produced an object that was not valid. It must result in a command
name, a script block, or a CommandInfo object.
At line:1 char:566
+ ...  (Should-Exclude $_.Path) }; foreach ($File in $Files) { & '' -mode=d ...
+                                                                ~~
    + CategoryInfo          : InvalidOperation: (:String) [], RuntimeException
    + FullyQualifiedErrorId : BadExpression
The expression after '&' in a pipeline element produced an object that was not valid. It must result in a command
[...]
```

The present change simply consists in giving a second try with
`$RUNFILES_MANIFEST_FILE` and failing fast should both be absent.